Allow backups to be restored to a larger cluster
This commit introduces the ability to restore a backup to a cluster that
has a different number of segments than the cluster which was originally
backed up.  Restoring to a larger cluster (one with more segments than
the original cluster) is fully supported; the framework for restoring to
a smaller cluster is in place, but is not yet fully functional, so that
has been disabled for now.

The general methodology for "resize" restores is as follows:

- Restoring to a larger cluster loads the data for a given segment in
the backup set to the segment with the corresponding content ID, then
redistributes the data.

- Restoring to a smaller cluster restores segment data in a round-robin
manner, where segments with a content ID greater than the highest one in
the destination cluster "wrap around" to content 0, restoring one batch
of segments at a time until all the data is loaded, then redistributes
the data.

- Any cases where an N-to-N restore is performed (both the default case for
a normal restore as well as batches within a larger-to-smaller restore) use
COPY ON SEGMENT to load the data; other cases use goroutines to load data
through individual COPY commands executed on each segment in parallel.

The --resize-cluster flag is added with this commit, and must be passed if
restoring to a differently-sized cluster in order to avoid any accidental
attempts to restore to a cluster of a different size.  It also adds a new
SegmentCount attribute to the configuration file, which must be present in
order to perform a resize restore; backups taken with an older version of
gpbackup may be restored with this flag if the parameter is manually added
to the config file manually, but taking a new backup is still recommended.